feat [#217] 마지막 사용자 활동 properties 추가
